Indeed Web Scrapper
Added the Indeed Web Scrapper from a private project. The Indeed Web Scrapper is able to scrape all the job postings from a given Indeed page (url is passed in, hard coded) and is able to get the Title of the job, the company name, the location, time posted and the indeed link. This data is stored into an excel file with the current date of running (Excel File Name: Job_Postings_for_D_M_Y)